Moving SMT into block-production directory
