+++ packages/core/src/app/block-production/index.ts
@@ -1,5 +1,6 @@
 export * from './block-db'
 export * from './block-manager'
+export * from './merkle-tree'
 export * from './merkle-interval-tree'
 export * from './state-interval-tree'
 export * from './plasma-block-tree'
