+++ packages/core/src/app/utils/index.ts
@@ -10,7 +10,6 @@ export const abi = new ethers.utils.AbiCoder()
 export * from './buffer'
 export * from './crypto'
 export * from './equals'
-export * from './merkle-tree'
 export * from './misc'
 export * from '../../types/number'
 export * from './range'
